Adds webpack, Creates components
