+++ public/index.html
@@ -40,6 +40,6 @@
       </section>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
