+++ src/components/load-more.js
@@ -1,0 +1,3 @@
+export const getLoadMoreBtn = () => {
+  return ` <button class="load-more" type="button">load more</button>`;
+};
