+++ src/utils.js
@@ -1,0 +1,3 @@
+export const renderComponent = (container, layoutTemplate, place) => {
+  container.insertAdjacentHTML(place, layoutTemplate);
+};
