+++ src/variables.js
@@ -1,0 +1,7 @@
+export const CARD_COLORS = {
+  black: `card--black`,
+  pink: `card--pink`,
+  yellow: `card--yellow`,
+  blue: `card--blue`,
+  green: `card--green`,
+};
